changed output file again2

+++ src/test_kernel_dma_invert_multiple_image.cpp
@@ -335,7 +335,7 @@ int main()
 		//char * temp = "";
 		for(int i = 0; i < LENGTH_OUTPUT; i++){
 			//sprintf(temp,"%d",finalResultBuff[i]);
-			Result_File << finalResultBuff[i] << "\n";
+			Result_File << (int)finalResultBuff[i] << "\n";
 		}
 		Result_File.close();
 		printf("\nSaved result to file!\n", TRIES_N);
